+++ CHANGELOG.md
@@ -49,6 +49,8 @@ See [RELEASE](./RELEASE.md) for workflow instructions.
 * [#5562](https://github.com/spacemeshos/go-spacemesh/pull/5562) Add streaming mode for fetcher. This should lessen
   GC pressure during sync
 
+* [#5718](https://github.com/spacemeshos/go-spacemesh/pull/5718) Sync malfeasance proofs continuously.
+
 ## Release v1.4.0
 
 ### Upgrade information